refactor(web): move location of UserBadges
